Add check_bucket_access().

+++ src/erlcloud_s3.erl
@@ -5,6 +5,7 @@
 -export([new/2, new/3, new/4,
          configure/2, configure/3, configure/4, configure/5,
          create_bucket/1, create_bucket/2, create_bucket/3, create_bucket/4,
+         check_bucket_access/1, check_bucket_access/2,
          delete_bucket/1, delete_bucket/2,
          get_bucket_attribute/2, get_bucket_attribute/3,
          list_buckets/0, list_buckets/1,
@@ -251,6 +252,23 @@ delete_bucket(BucketName) ->
 delete_bucket(BucketName, Config)
   when is_list(BucketName) ->
     s3_simple_request(Config, delete, BucketName, "/", "", [], <<>>, []).
+
+-spec check_bucket_access(string()) -> ok | {error, any()}.
+
+check_bucket_access(BucketName) 
+  when is_list(BucketName) ->
+    check_bucket_access(BucketName, default_config()).
+
+-spec check_bucket_access(string(), aws_config()) -> ok | {error, any()}.
+
+check_bucket_access(BucketName, Config)
+  when is_list(BucketName), is_record(Config, aws_config) ->
+    case s3_request2(Config, head, BucketName, "/", "", [], <<>>, []) of
+        {ok, {_Headers, _Body}} ->
+            ok;
+        Error ->
+            Error
+    end.
 
 
 -spec delete_objects_batch(string(), list()) -> no_return().
